casts enum to fix intel warning

+++ src/ref_cell.c
@@ -586,7 +586,7 @@ REF_STATUS ref_cell_inspect(REF_CELL ref_cell) {
 
 REF_STATUS ref_cell_tattle(REF_CELL ref_cell, REF_INT cell) {
   REF_INT node;
-  printf("cell %d %d:", ref_cell_type(ref_cell), cell);
+  printf("cell %d %d:", (int)ref_cell_type(ref_cell), cell);
   for (node = 0; node < ref_cell_size_per(ref_cell); node++) {
     printf(" %d", ref_cell_c2n(ref_cell, node, cell));
   }
